:bug: fix type mismatch (`uint8` to `uint32`)
uint8 と uint32 を間違えました

+++ models/image.go
@@ -1,7 +1,7 @@
 package models
 
 type ClubImage struct {
-	ImageID uint8  `gorm:"type:int unsigned;not null;primaryKey;autoIncrement"`
+	ImageID uint32 `gorm:"type:int unsigned;not null;primaryKey;autoIncrement"`
 	UUID    string `gorm:"type:char(36);not null"`
 	Path    string `gorm:"type:text;not null;unique"`
 }
