+++ models/schedule.go
@@ -1,7 +1,7 @@
 package models
 
 type ClubSchedule struct {
-	ScheduleID uint8  `gorm:"type:int unsigned;not null;primaryKey;autoIncrement"`
+	ScheduleID uint32 `gorm:"type:int unsigned;not null;primaryKey;autoIncrement"`
 	UUID       string `gorm:"type:char(36);not null"`
 	Month      uint8  `gorm:"type:tinyint unsigned;not null;unique"`
 	Schedule   string `gorm:"type:text;not null;unique"`
